requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
coloredlogs==10.0
humanfriendly==4.18
idna==2.8
praw==6.3.1
prawcore==1.0.1
python-Levenshtein==0.12.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
